+++ packages/infra/src/api/routing/middleware/generic-middleware.ts
@@ -40,6 +40,9 @@ export type DynamicMiddlewareMaker<RequestContext extends Record<string, RPCCont
 
 export namespace GenericMiddlewareMaker {
   export type ApplyServices<A extends TagClassAny, R> = Exclude<R, Provided<A>> | Required<A>
+  export type ApplyManyServices<A extends NonEmptyReadonlyArray<TagClassAny>, R> =
+    | Exclude<R, { [K in keyof A]: Provided<A[K]> }[number]>
+    | { [K in keyof A]: Required<A[K]> }[number]
   export type Provided<T> = T extends TagClassAny
     ? T extends { provides: Context.Tag<any, any> } ? Context.Tag.Identifier<T["provides"]>
     : T extends { provides: ContextRepr } ? ContextRepr.Identifier<T["provides"]>
