+++ packages/infra/src/api/routing/middleware/middleware-api.ts
@@ -1,4 +1,4 @@
-import { type NonEmptyReadonlyArray } from "effect-app"
+import { Array, Either, type NonEmptyArray } from "effect-app"
 import { type RPCContextMap } from "effect-app/client"
 import { type DynamicMiddlewareMaker, type GenericMiddlewareMaker, makeMiddleware, type makeMiddlewareBasic, type RequestContextMapProvider } from "../../routing.js"
 
@@ -9,39 +9,43 @@ export const contextMap = <RequestContextMap>() => <K extends keyof RequestConte
   settings: null as any as RequestContextMap[typeof a]
 })
 
-type DynamicMiddlewareMakerrsss<
+export interface MiddlewareM<
   RequestContext extends Record<string, RPCContextMap.Any>,
   Provided extends keyof RequestContext,
   Middlewares extends ReadonlyArray<GenericMiddlewareMaker>,
-  DynamicMiddlewareProviders extends RequestContextMapProvider<RequestContext>,
+  DynamicMiddlewareProviders,
   // out MiddlewareR = never
   MiddlewareR = never
-> = keyof Omit<RequestContext, Provided> extends never ?
-    & ReturnType<typeof makeMiddlewareBasic<RequestContext, DynamicMiddlewareProviders, Middlewares>>
-    & {
-      middleware: <MW extends GenericMiddlewareMaker>(
-        mw: MW
-      ) => DynamicMiddlewareMakerrsss<
-        RequestContext,
-        Provided,
-        [...Middlewares, MW],
-        DynamicMiddlewareProviders,
-        GenericMiddlewareMaker.ApplyServices<MW, MiddlewareR>
-      >
-    }
-  : {
-    middleware: <MW extends GenericMiddlewareMaker>(
-      mw: MW
-    ) => DynamicMiddlewareMakerrsss<
-      RequestContext,
-      Provided,
-      Middlewares,
-      DynamicMiddlewareProviders,
-      GenericMiddlewareMaker.ApplyServices<MW, MiddlewareR>
-    >
-    addDynamicMiddleware: <MW extends NonEmptyReadonlyArray<DynamicMiddlewareMaker<RequestContext>>>(
-      ...middlewares: MW
-    ) => DynamicMiddlewareMakerrsss<
+> {
+  middleware<MW extends NonEmptyArray<GenericMiddlewareMaker>>(
+    ...mw: MW
+  ): DynamicMiddlewareMakerrsss<
+    RequestContext,
+    Provided,
+    [...Middlewares, ...MW],
+    DynamicMiddlewareProviders,
+    GenericMiddlewareMaker.ApplyManyServices<MW, MiddlewareR>
+  >
+}
+
+export interface MiddlewareDynamic<
+  RequestContext extends Record<string, RPCContextMap.Any>,
+  Provided extends keyof RequestContext,
+  Middlewares extends ReadonlyArray<GenericMiddlewareMaker>,
+  DynamicMiddlewareProviders,
+  out MiddlewareR
+> extends
+  MiddlewareM<
+    RequestContext,
+    Provided,
+    Middlewares,
+    DynamicMiddlewareProviders,
+    MiddlewareR
+  >
+{
+  middleware<MW extends NonEmptyArray<DynamicMiddlewareMaker<RequestContext>> | NonEmptyArray<GenericMiddlewareMaker>>(
+    ...mw: MW
+  ): MW extends NonEmptyArray<DynamicMiddlewareMaker<RequestContext>> ? DynamicMiddlewareMakerrsss<
       RequestContext,
       Provided | MW[number]["dynamic"]["key"],
       Middlewares,
@@ -49,27 +53,101 @@ type DynamicMiddlewareMakerrsss<
       & {
         [K in keyof MW as MW[K] extends DynamicMiddlewareMaker<RequestContext> ? MW[K]["dynamic"]["key"] : never]: MW[K]
       },
-      MiddlewareR
-    > // GenericMiddlewareMaker.ApplyServices<MW, MiddlewareR>
-  }
+      GenericMiddlewareMaker.ApplyManyServices<MW, MiddlewareR>
+    >
+    : DynamicMiddlewareMakerrsss<
+      RequestContext,
+      Provided,
+      [...Middlewares, ...MW],
+      DynamicMiddlewareProviders,
+      GenericMiddlewareMaker.ApplyManyServices<MW, MiddlewareR>
+    >
+  // addDynamicMiddleware: <MW extends NonEmptyReadonlyArray<DynamicMiddlewareMaker<RequestContext>>>(
+  //   ...middlewares: MW
+  // ) => DynamicMiddlewareMakerrsss<
+  //   RequestContext,
+  //   Provided | MW[number]["dynamic"]["key"],
+  //   Middlewares,
+  //   & DynamicMiddlewareProviders
+  //   & {
+  //     [K in keyof MW as MW[K] extends DynamicMiddlewareMaker<RequestContext> ? MW[K]["dynamic"]["key"] : never]: MW[K]
+  //   },
+  //   MiddlewareR // TODO GenericMiddlewareMaker.ApplyServices<MW, MiddlewareR>
+  // > // GenericMiddlewareMaker.ApplyServices<MW, MiddlewareR>
+}
+
+type GetDynamicMiddleware<T, RequestContext extends Record<string, RPCContextMap.Any>> = T extends
+  RequestContextMapProvider<RequestContext> ? T : never
+
+type DynamicMiddlewareMakerrsss<
+  RequestContext extends Record<string, RPCContextMap.Any>,
+  Provided extends keyof RequestContext = never,
+  Middlewares extends ReadonlyArray<GenericMiddlewareMaker> = [],
+  DynamicMiddlewareProviders = unknown,
+  MiddlewareR = never
+> = keyof Omit<RequestContext, Provided> extends never ? [MiddlewareR] extends [never] ?
+      & ReturnType<
+        typeof makeMiddlewareBasic<
+          RequestContext,
+          GetDynamicMiddleware<DynamicMiddlewareProviders, RequestContext>,
+          Middlewares
+        >
+      >
+      // & {
+      //   MiddlewareR: MiddlewareR
+      //   Provided: Provided
+      //   Middlewares: Middlewares
+      //   DynamicMiddlewareProviders: Simplify<DynamicMiddlewareProviders>
+      // }
+      & MiddlewareM<
+        RequestContext,
+        Provided,
+        Middlewares,
+        DynamicMiddlewareProviders,
+        MiddlewareR
+      >
+  : MiddlewareM<
+    RequestContext,
+    Provided,
+    Middlewares,
+    DynamicMiddlewareProviders,
+    MiddlewareR
+  >
+  : MiddlewareDynamic<
+    RequestContext,
+    Provided,
+    Middlewares,
+    DynamicMiddlewareProviders,
+    MiddlewareR
+  >
 
 export const makeNewMiddleware: <
   RequestContextMap extends Record<string, RPCContextMap.Any>
->() => <Middlewares extends ReadonlyArray<GenericMiddlewareMaker>>(
-  ...genericMiddlewares: Middlewares
-) => DynamicMiddlewareMakerrsss<RequestContextMap, never, Middlewares, never> = () => (...genericMiddlewares) => {
-  const dynamicMiddlewares: Record<string, any> = {} as any
+>() => DynamicMiddlewareMakerrsss<RequestContextMap> = () => {
   const make = makeMiddleware<any>()
+  let capturedMiddlewares: (DynamicMiddlewareMaker<any> | GenericMiddlewareMaker)[] = []
   const it = {
-    middleware: <MW extends GenericMiddlewareMaker>(mw: MW) => {
-      genericMiddlewares = [...genericMiddlewares, mw] as any
-      return Object.assign(make({ genericMiddlewares, dynamicMiddlewares }), it)
-    },
-    addDynamicMiddleware: (...middlewares: any[]) => {
-      for (const a of middlewares) {
-        console.log("Adding dynamic middleware", a, a.dynamic.key, Object.keys(a))
-        dynamicMiddlewares[a.dynamic.key] = a
+    middleware: (...middlewares: any[]) => {
+      for (const mw of middlewares) {
+        capturedMiddlewares = [mw, ...capturedMiddlewares]
+        if (mw.dynamic) {
+          console.log("Adding dynamic middleware", mw.key, mw.dynamic.key)
+        } else {
+          console.log("Adding generic middleware", mw.key)
+        }
       }
+      const [genericMiddlewares, dyn] = Array.partitionMap(
+        capturedMiddlewares,
+        (mw) =>
+          "dynamic" in mw && mw.dynamic
+            ? Either.right(mw as DynamicMiddlewareMaker<any>)
+            : Either.left(mw as GenericMiddlewareMaker)
+      )
+      const dynamicMiddlewares = dyn.reduce(
+        (prev, cur) => ({ ...prev, [cur.dynamic.key]: cur }),
+        {} as Record<string, any>
+      )
+      // TODO: support dynamic and generic intertwined. treat them as one
       return Object.assign(make({ genericMiddlewares, dynamicMiddlewares }), it)
     }
   }
